refactor: colocate cart page components with route
Move 7 page-only cart components from components/cart/ into
app/cart/components/ — colocated with the route that uses them.

components/cart/ now contains only global infrastructure:
context, context-sync, actions, constants, optimistic-info,
and the overlay components (always present in layout).

Moved to app/cart/components/:
- header, items-section, item-row, summary
- empty-cart, skeletons, upsell-recommendations
