+++ apps/template/app/cart/components/skeletons.tsx
